refactor(errors): rename shared error codes module
