+++ src/chat-message-handler-stream.ts
@@ -1,7 +1,7 @@
 import { type ChatRow, createRow } from "./chat-contract";
 import { palette } from "./palette";
 import { createId } from "./short-id";
-import { LIFECYCLE_ERROR_CODES } from "./tool-error-codes";
+import { LIFECYCLE_ERROR_CODES } from "./error-codes";
 import { createToolOutputState, type ToolOutput } from "./tool-output-content";
 
 export type MessageStreamState = {
