+++ src/tool-execution.ts
@@ -1,6 +1,6 @@
 import { invariant } from "./assert";
 import { createId } from "./short-id";
-import { ERROR_KINDS, LIFECYCLE_ERROR_CODES, ToolError } from "./tool-error-codes";
+import { ERROR_KINDS, LIFECYCLE_ERROR_CODES, ToolError } from "./error-codes";
 import { recordCall, runGuards, type SessionContext } from "./tool-guards";
 
 function withTimeout<T>(task: () => Promise<T>, timeoutMs: number, toolId: string): Promise<T> {
